Bugfix: PrecompiledTransactionResult should correctly return its populated SideEffects
